azure-function/requirements.txt
azure-functions
openai>=1.0.0
azure-identity>=1.12.0
